MovePatternに定義されている移動関数のシグネチャを変更した。
元々は弾や敵にしか適用できなかった各関数を、get_position()メソッドを持っている型すべてにおいて扱えるようにしました。
また、それに合わせてメソッド名も汎用的に変更しています。

それと、相手の座標に向かって進む移動方法は引数に座標へのconst参照のみを取るようにしました。
